Move RaoUtil to RaoCommons and invert dependency
