+++ ra-optimisation/rao-commons/src/main/java/com/farao_community/farao/rao_commons/linear_optimisation/fillers/MaxMinMarginFiller.java
@@ -19,13 +19,13 @@ import com.google.ortools.linearsolver.MPVariable;
 import java.util.Optional;
 
 import static com.farao_community.farao.commons.Unit.MEGAWATT;
+import static com.farao_community.farao.rao_api.RaoParameters.DEFAULT_PST_PENALTY_COST;
 
 /**
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
  */
 public class MaxMinMarginFiller implements ProblemFiller {
-    public static final double DEFAULT_PST_PENALTY_COST = 0.01;
 
     private Unit unit;
     private double pstPenaltyCost;
