+++ ra-optimisation/search-tree-rao/src/main/java/com/farao_community/farao/search_tree_rao/Leaf.java
@@ -13,7 +13,7 @@ import com.farao_community.farao.data.crac_result_extensions.NetworkActionResult
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_commons.LoopFlowComputationService;
 import com.farao_community.farao.rao_commons.RaoData;
-import com.farao_community.farao.rao_api.RaoUtil;
+import com.farao_community.farao.rao_commons.RaoUtil;
 import com.farao_community.farao.rao_commons.linear_optimisation.iterating_linear_optimizer.IteratingLinearOptimizer;
 import com.farao_community.farao.rao_commons.SystematicSensitivityComputation;
 import com.powsybl.iidm.network.Network;
